+++ src/Himari/Prelude.hs
@@ -7,6 +7,7 @@ import Control.Monad.Primitive as Export
 import Data.Coerce as Export
 import Debug.Pretty.Simple as Export
 import Himari.Prelude.Aeson as Export
+import Himari.Prelude.Casing as Export
 import Himari.Prelude.Process as Export
 import Himari.Prelude.TypeLevel as Export
 import Himari.SafePrelude as Export